requirements_api.txt
# Requirements для API сервера
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
python-multipart>=0.0.6
websockets>=12.0
pyyaml>=6.0
